requirements.txt
pydantic==2.5.0
PyPDF2==3.0.1
pytest==8.4.1
pathlib2==2.3.7
streamlit==1.28.0
python-dateutil==2.8.2
PyYAML==6.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0 